Include battery level in join; allow charging beyond starting level

+++ src/Mars.MissionControl/Player.cs
@@ -16,7 +16,7 @@ public record Player
     }
 
     public TimeSpan? WinningTime { get; set; }
-    public int BatteryLevel { get; init; }
+    public long BatteryLevel { get; init; }
     public PlayerToken Token { get; private set; }
     public string Name { get; private set; }
     public Location PerseveranceLocation { get; init; }
