+++ src/Mars.MissionControl/Records.cs
@@ -2,8 +2,8 @@
 
 public record Location(int X, int Y);
 public record Cell(Location Location, Difficulty Difficulty);
-public record JoinResult(PlayerToken Token, Location PlayerLocation, Orientation Orientation, int BatteryLevel, IEnumerable<Location> TargetLocations, IEnumerable<Cell> Neighbors, IEnumerable<LowResolutionCell> LowResolutionMap);
-public record MoveResult(Location Location, int BatteryLevel, Orientation Orientation, IEnumerable<Cell> Neighbors, string Message);
+public record JoinResult(PlayerToken Token, Location PlayerLocation, Orientation Orientation, long BatteryLevel, IEnumerable<Location> TargetLocations, IEnumerable<Cell> Neighbors, IEnumerable<LowResolutionCell> LowResolutionMap);
+public record MoveResult(Location Location, long BatteryLevel, Orientation Orientation, IEnumerable<Cell> Neighbors, string Message);
 public record IngenuityMoveResult(Location Location, int BatteryLevel, IEnumerable<Cell> Neighbors, string Message);
 
 public enum Orientation
